requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Multi-agent orchestration
crewai==0.28.8
crewai-tools==0.2.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langchain-community==0.0.12

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2
fakeredis==2.20.1  # For testing

# Task Queue
celery==5.3.4
flower==2.0.1  # Celery monitoring

# Object Storage
minio==7.2.0
boto3==1.34.10  # S3 compatibility

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
argon2-cffi==23.1.0

# Rate Limiting
slowapi==0.1.9
limits==3.7.0

# Neo4j (optional graph database)
neo4j==5.14.1

# AI/ML frameworks
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.1

# Vector stores
chromadb==0.4.18
faiss-cpu==1.7.4

# PDF processing
PyMuPDF==1.23.8
requests==2.31.0

# Presentation and report generation
python-pptx==0.6.23
markdown==3.5.1
weasyprint==60.1

# Paper APIs and web scraping
arxiv==2.1.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Evaluation metrics
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1

# Data processing
numpy==1.24.3
pandas==2.1.3

# Web UI
streamlit==1.28.2
plotly==5.18.0

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
ruff==0.1.6

# Utilities
tqdm==4.66.1
aiohttp==3.9.1
tenacity==8.2.3
python-multipart==0.0.6
pillow==10.1.0

# Graph and network analysis
networkx==3.2.1
community==1.0.0b1
python-louvain==0.16

# Graph embeddings and ML
scikit-learn==1.3.2
node2vec==0.4.6
stellargraph==1.2.1

# Causal inference
dowhy==0.11.1
causalnex==0.12.1
econml==0.15.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0
